poc(): try to make a POC with scala / akka to native compile
Akka seems tricky to build, the class AkkaHttpGraal contains a sample that
builds fine with sbt (https://github.com/vmencik/akka-graal-native)

Currently, the build do not work with it (but using gluon instead of graal for the time being).

Tried with the two JDK from the documentation, without luck, errors are not the same thought.